+++ lib/src/main/java/gdv/xport/feld/NumFeld.java
@@ -25,6 +25,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.math.BigDecimal;
+import java.math.RoundingMode;
 import java.text.DecimalFormat;
 import java.text.NumberFormat;
 import java.util.List;
@@ -278,7 +279,7 @@ public class NumFeld extends Feld {
      * @since 5.0
      */
     public void setInhalt(BigDecimal n) {
-        setInhalt(n.movePointRight(this.nachkommastellen).setScale(0, ROUND_HALF_UP).toString());
+        setInhalt(n.movePointRight(this.nachkommastellen).setScale(0, RoundingMode.HALF_UP).toString());
     }
 
     /* (non-Javadoc)
@@ -290,6 +291,17 @@ public class NumFeld extends Feld {
         for (int i = 0; i < anzahlBytes; i++) {
             this.setInhalt('0', i);
         }
+    }
+
+    @Override
+    protected String truncate(String s) {
+        if (s.startsWith("0")) {
+            return truncate(s.substring(1));
+        } else if (s.length() > getAnzahlBytes()) {
+            LOG.info("NumFeld {} wird auf {} Ziffern gekuerzt", getBezeichner(), getAnzahlBytes());
+            return StringUtils.repeat('9', getAnzahlBytes());
+        }
+        return s;
     }
 
     /**
@@ -353,7 +365,7 @@ public class NumFeld extends Feld {
     public BigDecimal add(BigDecimal summand) {
         BigDecimal summe = toBigDecimal().add(summand);
         setInhalt(summe);
-        return summe.setScale(nachkommastellen, ROUND_UP);
+        return summe.setScale(nachkommastellen, RoundingMode.UP);
     }
 
     /**